okmen

+++ Client/client.c
@@ -40,8 +40,8 @@ void recvLtopic(int sockfd){
 }
 
 void sendFile (int connfd, char file_name[256]) {
-	int file_size;
-	int remain_data;
+	long int file_size;
+	long int remain_data;
 	struct stat st;
 	char buffer[1024];
   char *ptr;
@@ -58,8 +58,8 @@ void sendFile (int connfd, char file_name[256]) {
 		stat(ptr, &st);
 		remain_data = st.st_size;
 		file_size = remain_data;
-		printf("File_size: %d\n", file_size);
-		write(connfd, &file_size, sizeof(int));
+		printf("File_size: %ld\n", file_size);
+		write(connfd, &file_size, sizeof(long int));
 		while(remain_data > 0) {
 			if(remain_data < 1024) {
 				fread(buffer, remain_data, 1, fs);
@@ -78,12 +78,12 @@ void sendFile (int connfd, char file_name[256]) {
 
 void receiveFile (int sockfd, char file_name[256]) {
 	char buffer[1024];
-	int file_size;
+	long int file_size;
 	file_name[strcspn(file_name, "\n")]=0;
 	char *ptr;
 	ptr = strtok(file_name, "&");
-	read(sockfd, &file_size, sizeof(int));
-	printf("file_size la %d\n", file_size);
+	read(sockfd, &file_size, sizeof(long int));
+	printf("file_size la %ld\n", file_size);
 
 	if(file_size == -1) {
 		printf("File not exists\n");
